content 脚本信息传递 & injected

+++ scripts/pack.js
@@ -5,6 +5,9 @@ const { execSync } = require("child_process");
 
 let files = fs.readdirSync("./build/scriptcat/src");
 files.forEach(val => {
+    if (val === '.gitignore') {
+        return;
+    }
     fs.unlinkSync("./build/scriptcat/src/" + val);
 });
